qa-slave-shield.cpp and test-session.cpp fixes for Mini and RET6 Edition.
There's a bit of a hack on the RET6 NR_DAC_PINS: due to the fact that
PA5 is hooked up to the built-in LED and DAC channel 2, I set
NR_DAC_PINS to 0 for the STM32F103RE (in libmaple.h), so that init()
doesn't dac_init() and mess with the LED.

Not too worried about it since this is just a beta board, but we'll
need to fix this in the design for the real version.

Also, BUT should be moved, since it steals Timer 8's channel 4 CC
connection.