Add DELAY_US_MULT, for use in generalizing delay_us().
The delay_us() implementation multiplies its specified delay target by
a fixed constant in order to turn it into a busy-loop.  This magic
number doesn't work properly when the clock configuration isn't the
same as a stock LeafLabs board.  Add DELAY_US_MULT to the MCU-specific
configuration in stm32.h in order to allow other chips to use delay_us().