test-session.cpp: Better analog noise readings.
Update measure_adc_noise() to actually use the Welford online
algorithm, instead of accumulating data in an array on the stack.
This allows us to increase the number of samples (to 1000).

Revised algorithm tested on host PC and compared (in Python) against
numpy with a list of 100 values in [0, 1) drawn using random.random().

Results (Python):

    >>> r = [random.random() for i in xrange(100)]
    >>> numpy.mean(r)
    0.50073064742634854
    >>> numpy.var(r)
    0.083726090293309297

Results (C++, x86 host PC):

     n: 100	mean: 0.500731	variance: 0.084572

So this algorithm for variance has some inaccuracies, but it appears
to be good to a couple of significant figures.