stm32f1: stm32.h: Put MCU section first, allow overrides.
Put the section defining MCU-specific values before the other
sections. Surround the density-specific defines with #ifndef/#endif
pairs. This allows any of the settings in the STM32F1 stm32.h to be
overridden on a per-MCU basis. That's hopefully useful to e.g. people
porting libmaple to STM32F100 MCUs, which have slower clocks.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>