STM32F1: Fix nvic_irq_disable_all() on some MCUs.
The current implementation only disables the first 64 IRQ lines. This
covers all the chips we currently support, but it'll be a nasty
surprise if anyone decides to add e.g. connectivity line MCUs (which
have more IRQs) in the future.  We already have the infrastructure to
fix it in a clean way, so we might as well do it now.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>