Add CREDITS file.
It's about time. This initial file is based on the output of

$ git log | grep Author | sort | uniq

It leaves out credit for patches that were later reverted or moved out
of libmaple. It also doesn't credit the many patches and other
bugfixes from users who've e.g. written in the forum describing their
changes, but didn't submit actual patches. Names aren't always
available in cases like those, but if the contributors come forward,
we'd be happy to add them.

Add note in HACKING about adding yourself to CREDITS.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>