stm32f1/rules.mk: update include path for assembler files stm32f1/performance: load stm32.h in order to have STM32_HIGH_DENSITY when necessary
STM32_HIGH_DENSITY was not defined in isrs.S and vector_table.S for a high density STM32. This was because stm32.h where the flag is now defined (in contrast to being defined on the command line in an earlier version of libmaple) was not included. This led to crashes when using one of the high density peripherals.

Updated the ASFLAGS to have the correct include path in rules.mk.
Include stm32.h in the assembler files.